tests: split integration tests into separate directories and add a dedicated konnect suite
