+++ test/integration/ko/cleaner.go
