+++ test/integration/ko/dataplane_bluegreen_test.go
@@ -25,10 +25,11 @@ import (
 	testutils "github.com/kong/kong-operator/v2/pkg/utils/test"
 	"github.com/kong/kong-operator/v2/test/helpers"
 	"github.com/kong/kong-operator/v2/test/helpers/eventually"
+	"github.com/kong/kong-operator/v2/test/integration"
 )
 
 func TestDataPlaneBlueGreenRollout(t *testing.T) {
-	if !blueGreenController {
+	if !integration.BlueGreenControllerEnabled {
 		t.Skipf("KONG_OPERATOR_BLUEGREEN_CONTROLLER not set, skipping")
 	}
 	const (
@@ -36,7 +37,9 @@ func TestDataPlaneBlueGreenRollout(t *testing.T) {
 		tickTime = 100 * time.Millisecond
 	)
 
-	namespace, cleaner := helpers.SetupTestEnv(t, GetCtx(), GetEnv())
+	ctx := t.Context()
+	namespace, cleaner := helpers.SetupTestEnv(t, ctx, integration.GetEnv())
+	clients := integration.GetClients()
 
 	t.Log("deploying dataplane resource with 1 replica")
 	dataplaneName := types.NamespacedName{
@@ -51,40 +54,40 @@ func TestDataPlaneBlueGreenRollout(t *testing.T) {
 		Spec: testBlueGreenDataPlaneSpec(),
 	}
 
-	dataplaneClient := GetClients().OperatorClient.GatewayOperatorV1beta1().DataPlanes(namespace.Name)
+	dataplaneClient := integration.GetClients().OperatorClient.GatewayOperatorV1beta1().DataPlanes(namespace.Name)
 
-	dataplane, err := dataplaneClient.Create(GetCtx(), dataplane, metav1.CreateOptions{})
+	dataplane, err := dataplaneClient.Create(ctx, dataplane, metav1.CreateOptions{})
 	require.NoError(t, err)
 	cleaner.Add(dataplane)
 
 	t.Log("verifying dataplane gets marked ready")
-	require.Eventually(t, testutils.DataPlaneIsReady(t, GetCtx(), dataplaneName, GetClients().OperatorClient), waitTime, tickTime)
+	require.Eventually(t, testutils.DataPlaneIsReady(t, ctx, dataplaneName, integration.GetClients().OperatorClient), waitTime, tickTime)
 
 	t.Run("before patching", func(t *testing.T) {
 		t.Log("verifying preview deployment managed by the dataplane is present")
-		require.Eventually(t, testutils.DataPlaneHasDeployment(t, GetCtx(), dataplaneName, nil, clients, dataplanePreviewDeploymentLabels()), waitTime, tickTime)
+		require.Eventually(t, testutils.DataPlaneHasDeployment(t, ctx, dataplaneName, nil, clients, dataplanePreviewDeploymentLabels()), waitTime, tickTime)
 
 		t.Run("preview Admin API service", func(t *testing.T) {
 			t.Log("verifying preview admin service managed by the dataplane is present")
-			require.Eventually(t, testutils.DataPlaneHasService(t, GetCtx(), dataplaneName, clients, dataplaneAdminPreviewServiceLabels()), waitTime, tickTime)
+			require.Eventually(t, testutils.DataPlaneHasService(t, ctx, dataplaneName, clients, dataplaneAdminPreviewServiceLabels()), waitTime, tickTime)
 
 			t.Log("verifying that preview admin service has no active endpoints by default")
-			adminServices := testutils.MustListDataPlaneServices(t, GetCtx(), dataplane, GetClients().MgrClient, dataplaneAdminPreviewServiceLabels())
+			adminServices := testutils.MustListDataPlaneServices(t, ctx, dataplane, integration.GetClients().MgrClient, dataplaneAdminPreviewServiceLabels())
 			require.Len(t, adminServices, 1)
 			adminSvcNN := client.ObjectKeyFromObject(&adminServices[0])
-			require.Eventually(t, testutils.DataPlaneServiceHasNActiveEndpoints(t, GetCtx(), adminSvcNN, clients, 0), waitTime, tickTime,
+			require.Eventually(t, testutils.DataPlaneServiceHasNActiveEndpoints(t, ctx, adminSvcNN, clients, 0), waitTime, tickTime,
 				"with default rollout resource plan for DataPlane, the preview Admin Service shouldn't get an active endpoint")
 		})
 
 		t.Run("preview ingress service", func(t *testing.T) {
 			t.Log("verifying preview ingress service managed by the dataplane is present")
-			require.Eventually(t, testutils.DataPlaneHasService(t, GetCtx(), dataplaneName, clients, dataplaneIngressPreviewServiceLabels()), waitTime, tickTime)
+			require.Eventually(t, testutils.DataPlaneHasService(t, ctx, dataplaneName, clients, dataplaneIngressPreviewServiceLabels()), waitTime, tickTime)
 
 			t.Log("verifying that preview ingress service has no active endpoints by default")
-			ingressServices := testutils.MustListDataPlaneServices(t, GetCtx(), dataplane, GetClients().MgrClient, dataplaneIngressPreviewServiceLabels())
+			ingressServices := testutils.MustListDataPlaneServices(t, ctx, dataplane, integration.GetClients().MgrClient, dataplaneIngressPreviewServiceLabels())
 			require.Len(t, ingressServices, 1)
 			ingressSvcNN := client.ObjectKeyFromObject(&ingressServices[0])
-			require.Eventually(t, testutils.DataPlaneServiceHasNActiveEndpoints(t, GetCtx(), ingressSvcNN, clients, 0), waitTime, tickTime,
+			require.Eventually(t, testutils.DataPlaneServiceHasNActiveEndpoints(t, ctx, ingressSvcNN, clients, 0), waitTime, tickTime,
 				"with default rollout resource plan for DataPlane, the preview ingress Service shouldn't get an active endpoint")
 		})
 	})
@@ -92,47 +95,47 @@ func TestDataPlaneBlueGreenRollout(t *testing.T) {
 	dataplaneImageToPatch := helpers.GetDefaultDataPlaneBaseImage() + ":3.4"
 
 	t.Run("after patching", func(t *testing.T) {
-		patchDataPlaneImage(GetCtx(), t, dataplane, GetClients().MgrClient, dataplaneImageToPatch)
+		patchDataPlaneImage(ctx, t, dataplane, integration.GetClients().MgrClient, dataplaneImageToPatch)
 
 		t.Log("verifying preview deployment managed by the dataplane is present and has AvailableReplicas")
-		require.Eventually(t, testutils.DataPlaneHasActiveDeployment(t, GetCtx(), dataplaneName, &appsv1.Deployment{}, dataplanePreviewDeploymentLabels(), clients), waitTime, tickTime)
+		require.Eventually(t, testutils.DataPlaneHasActiveDeployment(t, ctx, dataplaneName, &appsv1.Deployment{}, dataplanePreviewDeploymentLabels(), clients), waitTime, tickTime)
 
 		t.Run("preview Admin API service", func(t *testing.T) {
 			t.Log("verifying preview admin service managed by the dataplane has an active endpoint")
-			require.Eventually(t, testutils.DataPlaneHasService(t, GetCtx(), dataplaneName, clients, dataplaneAdminPreviewServiceLabels()), waitTime, tickTime)
+			require.Eventually(t, testutils.DataPlaneHasService(t, ctx, dataplaneName, clients, dataplaneAdminPreviewServiceLabels()), waitTime, tickTime)
 
 			t.Log("verifying that preview admin service has an active endpoint")
-			adminServices := testutils.MustListDataPlaneServices(t, GetCtx(), dataplane, GetClients().MgrClient, dataplaneAdminPreviewServiceLabels())
+			adminServices := testutils.MustListDataPlaneServices(t, ctx, dataplane, integration.GetClients().MgrClient, dataplaneAdminPreviewServiceLabels())
 			require.Len(t, adminServices, 1)
 			adminSvcNN := client.ObjectKeyFromObject(&adminServices[0])
-			require.Eventually(t, testutils.DataPlaneServiceHasNActiveEndpoints(t, GetCtx(), adminSvcNN, clients, 1), waitTime, tickTime,
+			require.Eventually(t, testutils.DataPlaneServiceHasNActiveEndpoints(t, ctx, adminSvcNN, clients, 1), waitTime, tickTime,
 				"with default rollout resource plan for DataPlane, the preview Admin Service should get an active endpoint")
 		})
 
 		t.Run("preview ingress service", func(t *testing.T) {
 			t.Log("verifying preview ingress service managed by the dataplane has an active endpoint")
-			require.Eventually(t, testutils.DataPlaneHasService(t, GetCtx(), dataplaneName, clients, dataplaneIngressPreviewServiceLabels()), waitTime, tickTime)
+			require.Eventually(t, testutils.DataPlaneHasService(t, ctx, dataplaneName, clients, dataplaneIngressPreviewServiceLabels()), waitTime, tickTime)
 
 			t.Log("verifying that preview ingress service has an active endpoint")
-			ingressServices := testutils.MustListDataPlaneServices(t, GetCtx(), dataplane, GetClients().MgrClient, dataplaneIngressPreviewServiceLabels())
+			ingressServices := testutils.MustListDataPlaneServices(t, ctx, dataplane, integration.GetClients().MgrClient, dataplaneIngressPreviewServiceLabels())
 			require.Len(t, ingressServices, 1)
 			ingressSvcNN := client.ObjectKeyFromObject(&ingressServices[0])
-			require.Eventually(t, testutils.DataPlaneServiceHasNActiveEndpoints(t, GetCtx(), ingressSvcNN, clients, 1), waitTime, tickTime,
+			require.Eventually(t, testutils.DataPlaneServiceHasNActiveEndpoints(t, ctx, ingressSvcNN, clients, 1), waitTime, tickTime,
 				"with default rollout resource plan for DataPlane, the preview ingress Service should get an active endpoint")
 		})
 
 		t.Run("live ingress service", func(t *testing.T) {
 			t.Log("verifying that live ingress service managed by the dataplane is available")
 			var liveIngressService corev1.Service
-			require.Eventually(t, testutils.DataPlaneHasActiveService(t, GetCtx(), dataplaneName, &liveIngressService, clients, dataplaneIngressLiveServiceLabels()), waitTime, tickTime)
+			require.Eventually(t, testutils.DataPlaneHasActiveService(t, ctx, dataplaneName, &liveIngressService, clients, dataplaneIngressLiveServiceLabels()), waitTime, tickTime)
 
-			require.Eventually(t, testutils.DataPlaneServiceHasNActiveEndpoints(t, GetCtx(), client.ObjectKeyFromObject(&liveIngressService), clients, 1), waitTime, tickTime,
+			require.Eventually(t, testutils.DataPlaneServiceHasNActiveEndpoints(t, ctx, client.ObjectKeyFromObject(&liveIngressService), clients, 1), waitTime, tickTime,
 				"live ingress Service should always have an active endpoint")
 		})
 
 		t.Run("live deployment", func(t *testing.T) {
 			t.Log("verifying live deployment managed by the dataplane is present and has an available replica")
-			require.Eventually(t, testutils.DataPlaneHasActiveDeployment(t, GetCtx(), dataplaneName, &appsv1.Deployment{}, dataplaneLiveDeploymentLabels(), clients), waitTime, tickTime)
+			require.Eventually(t, testutils.DataPlaneHasActiveDeployment(t, ctx, dataplaneName, &appsv1.Deployment{}, dataplaneLiveDeploymentLabels(), clients), waitTime, tickTime)
 		})
 	})
 
@@ -156,14 +159,14 @@ func TestDataPlaneBlueGreenRollout(t *testing.T) {
 			Status: metav1.ConditionFalse,
 		})
 		require.Eventually(t,
-			testutils.DataPlanePredicate(t, GetCtx(), dataplaneName, isAwaitingPromotion, GetClients().OperatorClient),
+			testutils.DataPlanePredicate(t, ctx, dataplaneName, isAwaitingPromotion, integration.GetClients().OperatorClient),
 			waitTime, tickTime,
 		)
 	})
 
 	t.Run("after promotion", func(t *testing.T) {
 		t.Logf("patching DataPlane with promotion triggering annotation %s=%s", operatorv1beta1.DataPlanePromoteWhenReadyAnnotationKey, operatorv1beta1.DataPlanePromoteWhenReadyAnnotationTrue)
-		patchDataPlaneAnnotations(t, dataplane, GetClients().MgrClient, map[string]string{
+		patchDataPlaneAnnotations(t, dataplane, integration.GetClients().MgrClient, map[string]string{
 			operatorv1beta1.DataPlanePromoteWhenReadyAnnotationKey: operatorv1beta1.DataPlanePromoteWhenReadyAnnotationTrue,
 		})
 
@@ -171,7 +174,7 @@ func TestDataPlaneBlueGreenRollout(t *testing.T) {
 			t.Log("verifying live deployment managed by the dataplane is present and has an available replica using the patched proxy image")
 
 			require.Eventually(t,
-				testutils.DataPlaneHasDeployment(t, GetCtx(), dataplaneName, nil, clients, dataplaneLiveDeploymentLabels(),
+				testutils.DataPlaneHasDeployment(t, ctx, dataplaneName, nil, clients, dataplaneLiveDeploymentLabels(),
 					func(d appsv1.Deployment) bool {
 						proxyContainer := k8sutils.GetPodContainerByName(&d.Spec.Template.Spec, consts.DataPlaneProxyContainerName)
 						return proxyContainer != nil && dataplaneImageToPatch == proxyContainer.Image
@@ -183,21 +186,21 @@ func TestDataPlaneBlueGreenRollout(t *testing.T) {
 		t.Run("live ingress service", func(t *testing.T) {
 			t.Log("verifying that live ingress service managed by the dataplane still has active endpoints")
 			var liveIngressService corev1.Service
-			require.Eventually(t, testutils.DataPlaneHasActiveService(t, GetCtx(), dataplaneName, &liveIngressService, clients, dataplaneIngressLiveServiceLabels()), waitTime, tickTime)
+			require.Eventually(t, testutils.DataPlaneHasActiveService(t, ctx, dataplaneName, &liveIngressService, clients, dataplaneIngressLiveServiceLabels()), waitTime, tickTime)
 			require.NotNil(t, liveIngressService)
 
-			require.Eventually(t, testutils.DataPlaneServiceHasNActiveEndpoints(t, GetCtx(), client.ObjectKeyFromObject(&liveIngressService), clients, 1), waitTime, tickTime,
+			require.Eventually(t, testutils.DataPlaneServiceHasNActiveEndpoints(t, ctx, client.ObjectKeyFromObject(&liveIngressService), clients, 1), waitTime, tickTime,
 				"live ingress Service should always have an active endpoint")
 		})
 
 		t.Run(fmt.Sprintf("%s annotation is cleared from DataPlane", operatorv1beta1.DataPlanePromoteWhenReadyAnnotationKey), func(t *testing.T) {
 			require.Eventually(t,
-				testutils.DataPlanePredicate(t, GetCtx(), dataplaneName,
+				testutils.DataPlanePredicate(t, ctx, dataplaneName,
 					func(dataplane *operatorv1beta1.DataPlane) bool {
 						_, ok := dataplane.Annotations[operatorv1beta1.DataPlanePromoteWhenReadyAnnotationKey]
 						return !ok
 					},
-					GetClients().OperatorClient,
+					integration.GetClients().OperatorClient,
 				),
 				waitTime, tickTime,
 			)
@@ -208,14 +211,14 @@ func TestDataPlaneBlueGreenRollout(t *testing.T) {
 		t.Logf("patching DataPlane by removing the rollout strategy")
 		old := dataplane.DeepCopy()
 		dataplane.Spec.Deployment.Rollout = nil
-		require.NoError(t, GetClients().MgrClient.Patch(GetCtx(), dataplane, client.MergeFrom(old)))
+		require.NoError(t, integration.GetClients().MgrClient.Patch(ctx, dataplane, client.MergeFrom(old)))
 
 		t.Run("preview deployment", func(t *testing.T) {
 			t.Log("verifying that preview deployment managed by the dataplane is removed")
 
 			require.Eventually(t,
 				testutils.Not(
-					testutils.DataPlaneHasDeployment(t, GetCtx(), dataplaneName, nil, clients, dataplanePreviewDeploymentLabels())),
+					testutils.DataPlaneHasDeployment(t, ctx, dataplaneName, nil, clients, dataplanePreviewDeploymentLabels())),
 				waitTime, tickTime)
 		})
 
@@ -224,22 +227,22 @@ func TestDataPlaneBlueGreenRollout(t *testing.T) {
 			var previewIngressService corev1.Service
 			require.Eventually(t,
 				testutils.Not(
-					testutils.DataPlaneHasActiveService(t, GetCtx(), dataplaneName, &previewIngressService, clients, dataplaneIngressPreviewServiceLabels())),
+					testutils.DataPlaneHasActiveService(t, ctx, dataplaneName, &previewIngressService, clients, dataplaneIngressPreviewServiceLabels())),
 				waitTime, tickTime)
 		})
 
 		t.Run("dataplane status rollout should be cleared", func(t *testing.T) {
 			require.Eventually(t,
-				testutils.DataPlanePredicate(t, GetCtx(), dataplaneName, func(dataplane *operatorv1beta1.DataPlane) bool {
+				testutils.DataPlanePredicate(t, ctx, dataplaneName, func(dataplane *operatorv1beta1.DataPlane) bool {
 					return dataplane.Status.RolloutStatus == nil
-				}, GetClients().OperatorClient),
+				}, integration.GetClients().OperatorClient),
 				waitTime, tickTime)
 		})
 	})
 }
 
 func TestDataPlaneBlueGreenHorizontalScaling(t *testing.T) {
-	if !blueGreenController {
+	if !integration.BlueGreenControllerEnabled {
 		t.Skipf("KONG_OPERATOR_BLUEGREEN_CONTROLLER not set, skipping")
 	}
 	const (
@@ -247,7 +250,9 @@ func TestDataPlaneBlueGreenHorizontalScaling(t *testing.T) {
 		tickTime = 100 * time.Millisecond
 	)
 
-	namespace, cleaner := helpers.SetupTestEnv(t, GetCtx(), GetEnv())
+	ctx := t.Context()
+	clients := integration.GetClients()
+	namespace, cleaner := helpers.SetupTestEnv(t, ctx, integration.GetEnv())
 
 	t.Log("deploying a dataplane")
 	dataplaneName := types.NamespacedName{
@@ -281,23 +286,23 @@ func TestDataPlaneBlueGreenHorizontalScaling(t *testing.T) {
 		},
 	}
 
-	dataplaneClient := GetClients().OperatorClient.GatewayOperatorV1beta1().DataPlanes(namespace.Name)
+	dataplaneClient := integration.GetClients().OperatorClient.GatewayOperatorV1beta1().DataPlanes(namespace.Name)
 
-	dataplane, err := dataplaneClient.Create(GetCtx(), dataplane, metav1.CreateOptions{})
+	dataplane, err := dataplaneClient.Create(ctx, dataplane, metav1.CreateOptions{})
 	require.NoError(t, err)
 	cleaner.Add(dataplane)
 
 	t.Logf("verifying DataPlane %s gets marked ready", dataplane.Name)
-	require.Eventually(t, testutils.DataPlaneIsReady(t, GetCtx(), dataplaneName, GetClients().OperatorClient), waitTime, tickTime)
+	require.Eventually(t, testutils.DataPlaneIsReady(t, ctx, dataplaneName, integration.GetClients().OperatorClient), waitTime, tickTime)
 
 	dataplaneImageToPatch := helpers.GetDefaultDataPlaneBaseImage() + ":3.4"
-	patchDataPlaneImage(GetCtx(), t, dataplane, GetClients().MgrClient, dataplaneImageToPatch)
+	patchDataPlaneImage(ctx, t, dataplane, integration.GetClients().MgrClient, dataplaneImageToPatch)
 	var previewDeployment appsv1.Deployment
-	require.Eventually(t, testutils.DataPlaneHasActiveDeployment(t, GetCtx(), dataplaneName, &previewDeployment, dataplanePreviewDeploymentLabels(), clients), waitTime, tickTime)
+	require.Eventually(t, testutils.DataPlaneHasActiveDeployment(t, ctx, dataplaneName, &previewDeployment, dataplanePreviewDeploymentLabels(), clients), waitTime, tickTime)
 
 	t.Logf("checking if DataPlane %s has an HPA", dataplane.Name)
 	var hpa autoscalingv2.HorizontalPodAutoscaler
-	require.Eventually(t, testutils.DataPlaneHasHPA(t, GetCtx(), dataplane, &hpa, clients),
+	require.Eventually(t, testutils.DataPlaneHasHPA(t, ctx, dataplane, &hpa, clients),
 		waitTime, tickTime, "HPA should be created for DataPlane %s", dataplane.Name)
 	require.NotNil(t, hpa)
 	require.NotNil(t, hpa.Spec.MinReplicas)
@@ -309,14 +314,14 @@ func TestDataPlaneBlueGreenHorizontalScaling(t *testing.T) {
 	assert.Equal(t, int32(20), *hpa.Spec.Metrics[0].Resource.Target.AverageUtilization)
 
 	t.Logf("checking if the HPA %s is pointing to the correct Deployment", hpa.Name)
-	require.Eventually(t, testutils.HPAPredicate(t, GetCtx(), client.ObjectKeyFromObject(&hpa),
+	require.Eventually(t, testutils.HPAPredicate(t, ctx, client.ObjectKeyFromObject(&hpa),
 		func(hpa *autoscalingv2.HorizontalPodAutoscaler) bool {
 			return hpa.Spec.ScaleTargetRef.Name != previewDeployment.Name
-		}, GetClients().MgrClient),
+		}, integration.GetClients().MgrClient),
 		waitTime, tickTime, "HPA should not target the preview deployment")
 
 	var deployment2 appsv1.Deployment
-	require.Eventually(t, testutils.DataPlaneHasDeployment(t, GetCtx(), dataplaneName, &deployment2, clients,
+	require.Eventually(t, testutils.DataPlaneHasDeployment(t, ctx, dataplaneName, &deployment2, clients,
 		client.MatchingLabels{
 			consts.GatewayOperatorManagedByLabel: consts.DataPlaneManagedLabelValue,
 			consts.DataPlaneDeploymentStateLabel: consts.DataPlaneStateLabelValueLive,
@@ -326,19 +331,19 @@ func TestDataPlaneBlueGreenHorizontalScaling(t *testing.T) {
 		waitTime, tickTime, "HPA should target the live deployment")
 
 	t.Logf("patching DataPlane with promotion triggering annotation %s=%s", operatorv1beta1.DataPlanePromoteWhenReadyAnnotationKey, operatorv1beta1.DataPlanePromoteWhenReadyAnnotationTrue)
-	patchDataPlaneAnnotations(t, dataplane, GetClients().MgrClient, map[string]string{
+	patchDataPlaneAnnotations(t, dataplane, integration.GetClients().MgrClient, map[string]string{
 		operatorv1beta1.DataPlanePromoteWhenReadyAnnotationKey: operatorv1beta1.DataPlanePromoteWhenReadyAnnotationTrue,
 	})
 	t.Logf("HPA %s should now point to just promoted %s Deployment", hpa.Name, previewDeployment.Name)
-	require.Eventually(t, testutils.HPAPredicate(t, GetCtx(), client.ObjectKeyFromObject(&hpa),
+	require.Eventually(t, testutils.HPAPredicate(t, ctx, client.ObjectKeyFromObject(&hpa),
 		func(hpa *autoscalingv2.HorizontalPodAutoscaler) bool {
 			return hpa.Spec.ScaleTargetRef.Name == previewDeployment.Name
-		}, GetClients().MgrClient),
+		}, integration.GetClients().MgrClient),
 		waitTime, tickTime)
 }
 
 func TestDataPlaneBlueGreenResourcesNotDeletedUntilOwnerIsRemoved(t *testing.T) {
-	if !blueGreenController {
+	if !integration.BlueGreenControllerEnabled {
 		t.Skipf("KONG_OPERATOR_BLUEGREEN_CONTROLLER not set, skipping")
 	}
 	const (
@@ -346,7 +351,9 @@ func TestDataPlaneBlueGreenResourcesNotDeletedUntilOwnerIsRemoved(t *testing.T) 
 		tickTime = 100 * time.Millisecond
 	)
 
-	namespace, cleaner := helpers.SetupTestEnv(t, GetCtx(), GetEnv())
+	ctx := t.Context()
+	clients := integration.GetClients()
+	namespace, cleaner := helpers.SetupTestEnv(t, ctx, integration.GetEnv())
 
 	t.Log("deploying dataplane")
 	dataplane := &operatorv1beta1.DataPlane{
@@ -358,8 +365,8 @@ func TestDataPlaneBlueGreenResourcesNotDeletedUntilOwnerIsRemoved(t *testing.T) 
 	}
 	dataplaneName := client.ObjectKeyFromObject(dataplane)
 
-	dataplaneClient := GetClients().OperatorClient.GatewayOperatorV1beta1().DataPlanes(namespace.Name)
-	dataplane, err := dataplaneClient.Create(GetCtx(), dataplane, metav1.CreateOptions{})
+	dataplaneClient := integration.GetClients().OperatorClient.GatewayOperatorV1beta1().DataPlanes(namespace.Name)
+	dataplane, err := dataplaneClient.Create(ctx, dataplane, metav1.CreateOptions{})
 	require.NoError(t, err)
 	cleaner.Add(dataplane)
 
@@ -370,21 +377,21 @@ func TestDataPlaneBlueGreenResourcesNotDeletedUntilOwnerIsRemoved(t *testing.T) 
 		liveDeployment     = &appsv1.Deployment{}
 		liveTLSSecret      = &corev1.Secret{}
 	)
-	require.Eventually(t, testutils.DataPlaneHasActiveService(t, GetCtx(), dataplaneName, liveIngressService, clients, dataplaneIngressLiveServiceLabels()), waitTime, tickTime)
+	require.Eventually(t, testutils.DataPlaneHasActiveService(t, ctx, dataplaneName, liveIngressService, clients, dataplaneIngressLiveServiceLabels()), waitTime, tickTime)
 	require.NotNil(t, liveIngressService)
 
-	require.Eventually(t, testutils.DataPlaneHasActiveService(t, GetCtx(), dataplaneName, liveAdminService, clients, dataplaneAdminLiveServiceLabels()), waitTime, tickTime)
+	require.Eventually(t, testutils.DataPlaneHasActiveService(t, ctx, dataplaneName, liveAdminService, clients, dataplaneAdminLiveServiceLabels()), waitTime, tickTime)
 	require.NotNil(t, liveAdminService)
 
-	require.Eventually(t, testutils.DataPlaneHasActiveDeployment(t, GetCtx(), dataplaneName, liveDeployment, dataplaneLiveDeploymentLabels(), clients), waitTime, tickTime)
+	require.Eventually(t, testutils.DataPlaneHasActiveDeployment(t, ctx, dataplaneName, liveDeployment, dataplaneLiveDeploymentLabels(), clients), waitTime, tickTime)
 	require.NotNil(t, liveDeployment)
 
-	require.Eventually(t, testutils.DataPlaneHasServiceSecret(t, GetCtx(), dataplaneName, client.ObjectKeyFromObject(liveAdminService), liveTLSSecret, clients), waitTime, tickTime)
+	require.Eventually(t, testutils.DataPlaneHasServiceSecret(t, ctx, dataplaneName, client.ObjectKeyFromObject(liveAdminService), liveTLSSecret, clients), waitTime, tickTime)
 	require.NotNil(t, liveTLSSecret)
 
 	t.Log("patching dataplane with another dataplane image to trigger rollout")
 	dataplaneImageToPatch := helpers.GetDefaultDataPlaneBaseImage() + ":3.4"
-	patchDataPlaneImage(GetCtx(), t, dataplane, GetClients().MgrClient, dataplaneImageToPatch)
+	patchDataPlaneImage(ctx, t, dataplane, integration.GetClients().MgrClient, dataplaneImageToPatch)
 
 	t.Log("ensuring all preview dependent resources are created")
 	var (
@@ -394,16 +401,16 @@ func TestDataPlaneBlueGreenResourcesNotDeletedUntilOwnerIsRemoved(t *testing.T) 
 		previewTLSSecret      = &corev1.Secret{}
 	)
 
-	require.Eventually(t, testutils.DataPlaneHasActiveService(t, GetCtx(), dataplaneName, previewIngressService, clients, dataplaneIngressPreviewServiceLabels()), waitTime, tickTime)
+	require.Eventually(t, testutils.DataPlaneHasActiveService(t, ctx, dataplaneName, previewIngressService, clients, dataplaneIngressPreviewServiceLabels()), waitTime, tickTime)
 	require.NotNil(t, previewIngressService)
 
-	require.Eventually(t, testutils.DataPlaneHasActiveService(t, GetCtx(), dataplaneName, previewAdminService, clients, dataplaneAdminPreviewServiceLabels()), waitTime, tickTime)
+	require.Eventually(t, testutils.DataPlaneHasActiveService(t, ctx, dataplaneName, previewAdminService, clients, dataplaneAdminPreviewServiceLabels()), waitTime, tickTime)
 	require.NotNil(t, previewAdminService)
 
-	require.Eventually(t, testutils.DataPlaneHasActiveDeployment(t, GetCtx(), dataplaneName, previewDeployment, dataplanePreviewDeploymentLabels(), clients), waitTime, tickTime)
+	require.Eventually(t, testutils.DataPlaneHasActiveDeployment(t, ctx, dataplaneName, previewDeployment, dataplanePreviewDeploymentLabels(), clients), waitTime, tickTime)
 	require.NotNil(t, previewDeployment)
 
-	require.Eventually(t, testutils.DataPlaneHasServiceSecret(t, GetCtx(), dataplaneName, client.ObjectKeyFromObject(previewAdminService), previewTLSSecret, clients), waitTime, tickTime)
+	require.Eventually(t, testutils.DataPlaneHasServiceSecret(t, ctx, dataplaneName, client.ObjectKeyFromObject(previewAdminService), previewTLSSecret, clients), waitTime, tickTime)
 	require.NotNil(t, previewTLSSecret)
 
 	dependentResources := []client.Object{
@@ -419,10 +426,10 @@ func TestDataPlaneBlueGreenResourcesNotDeletedUntilOwnerIsRemoved(t *testing.T) 
 
 	t.Log("ensuring dataplane owned resources after deletion are not immediately deleted")
 	for _, resource := range dependentResources {
-		require.NoError(t, GetClients().MgrClient.Delete(GetCtx(), resource))
+		require.NoError(t, integration.GetClients().MgrClient.Delete(ctx, resource))
 
 		require.Eventually(t, func() bool {
-			err := GetClients().MgrClient.Get(GetCtx(), client.ObjectKeyFromObject(resource), resource)
+			err := integration.GetClients().MgrClient.Get(ctx, client.ObjectKeyFromObject(resource), resource)
 			if err != nil {
 				t.Logf("error getting %T: %v", resource, err)
 				return false
@@ -438,9 +445,9 @@ func TestDataPlaneBlueGreenResourcesNotDeletedUntilOwnerIsRemoved(t *testing.T) 
 	}
 
 	t.Log("deleting dataplane and ensuring its owned resources are deleted after that")
-	require.NoError(t, GetClients().MgrClient.Delete(GetCtx(), dataplane))
+	require.NoError(t, integration.GetClients().MgrClient.Delete(ctx, dataplane))
 	for _, resource := range dependentResources {
-		eventually.WaitForObjectToNotExist(t, ctx, GetClients().MgrClient, resource, waitTime, tickTime,
+		eventually.WaitForObjectToNotExist(t, ctx, integration.GetClients().MgrClient, resource, waitTime, tickTime,
 			"should be deleted after dataplane deletion",
 		)
 	}
@@ -502,6 +509,7 @@ func patchDataPlaneImage(ctx context.Context, t *testing.T, dataplane *operatorv
 }
 
 func patchDataPlaneAnnotations(t *testing.T, dataplane *operatorv1beta1.DataPlane, cl client.Client, annotations map[string]string) {
+	ctx := t.Context()
 	oldDataPlane := dataplane.DeepCopy()
 	require.Len(t, dataplane.Spec.Deployment.PodTemplateSpec.Spec.Containers, 1)
 	if dataplane.Annotations == nil {
@@ -509,7 +517,7 @@ func patchDataPlaneAnnotations(t *testing.T, dataplane *operatorv1beta1.DataPlan
 	} else {
 		maps.Copy(dataplane.Annotations, annotations)
 	}
-	require.NoError(t, cl.Patch(GetCtx(), dataplane, client.MergeFrom(oldDataPlane)))
+	require.NoError(t, cl.Patch(ctx, dataplane, client.MergeFrom(oldDataPlane)))
 }
 
 func dataplaneAdminPreviewServiceLabels() client.MatchingLabels {
