+++ test/integration/ko/utils.go
@@ -1,0 +1,47 @@
+package integration
+
+import (
+	"context"
+	"fmt"
+	"net/url"
+
+	"github.com/kong/kubernetes-testing-framework/pkg/clusters"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+)
+
+// URLForService returns the URL for the given service and port.
+// It supports LoadBalancer and ClusterIP services.
+func URLForService(ctx context.Context, cluster clusters.Cluster, nsn types.NamespacedName, port int) (*url.URL, error) {
+	service, err := cluster.Client().CoreV1().Services(nsn.Namespace).Get(ctx, nsn.Name, metav1.GetOptions{})
+	if err != nil {
+		return nil, err
+	}
+
+	switch service.Spec.Type {
+	case corev1.ServiceTypeLoadBalancer:
+		if len(service.Status.LoadBalancer.Ingress) == 1 {
+			return url.Parse(fmt.Sprintf("http://%s:%d", service.Status.LoadBalancer.Ingress[0].IP, port))
+		}
+	default:
+		if service.Spec.ClusterIP != "" {
+			return url.Parse(fmt.Sprintf("http://%s:%d", service.Spec.ClusterIP, port))
+		}
+	}
+
+	return nil, fmt.Errorf("service %s has not yet been provisoned", service.Name)
+}
+
+// GetEnvValueFromByName returns the corresponding ValueFrom pointer of LAST item with given name.
+// returns nil if the name not appeared.
+func GetEnvValueFromByName(envs []corev1.EnvVar, name string) *corev1.EnvVarSource {
+	var valueFrom *corev1.EnvVarSource
+	for _, env := range envs {
+		if env.Name == name {
+			valueFrom = env.ValueFrom
+		}
+	}
+
+	return valueFrom
+}
