+++ test/integration/konnect/consts.go
@@ -1,0 +1,8 @@
+package konnect
+
+import "time"
+
+const (
+	waitTime = 60 * time.Second
+	tickTime = 1 * time.Second
+)
