+++ test/integration/konnect/konnect_entities_test.go
@@ -1,6 +1,7 @@
-package integration
+package konnect
 
 import (
+	"context"
 	"strings"
 	"testing"
 
@@ -24,17 +25,22 @@ import (
 	"github.com/kong/kong-operator/v2/test/helpers/conditions"
 	"github.com/kong/kong-operator/v2/test/helpers/deploy"
 	"github.com/kong/kong-operator/v2/test/helpers/eventually"
+	"github.com/kong/kong-operator/v2/test/helpers/object"
+	"github.com/kong/kong-operator/v2/test/integration"
 )
 
 func TestKonnectEntities(t *testing.T) {
+	ctx := t.Context()
+	cl := integration.GetClients().MgrClient
+
 	// A cleaner is created underneath anyway, and a whole namespace is deleted eventually.
 	// We can't use a cleaner to delete objects because it handles deletes in FIFO order and that won't work in this
 	// case: KonnectAPIAuthConfiguration shouldn't be deleted before any other object as that is required for others to
 	// complete their finalizer which is deleting a reflecting entity in Konnect. That's why we're only cleaning up a
-	// KonnectGatewayControlPlane and waiting for its deletion synchronously with deleteObjectAndWaitForDeletionFn to ensure it
+	// KonnectGatewayControlPlane and waiting for its deletion synchronously with delete.ObjectAndWaitForDeletionFn to ensure it
 	// was successfully deleted along with its children. The KonnectAPIAuthConfiguration is implicitly deleted along
 	// with the namespace.
-	ns, _ := helpers.SetupTestEnv(t, GetCtx(), GetEnv())
+	ns, _ := helpers.SetupTestEnv(t, ctx, integration.GetEnv())
 
 	// Let's generate a unique test ID that we can refer to in Konnect entities.
 	// Using only the first 8 characters of the UUID to keep the ID short enough for Konnect to accept it as a part
@@ -42,9 +48,9 @@ func TestKonnectEntities(t *testing.T) {
 	testID := uuid.NewString()[:8]
 	t.Logf("Running Konnect entities test with ID: %s", testID)
 
-	clientNamespaced := client.NewNamespacedClient(GetClients().MgrClient, ns.Name)
+	clientNamespaced := client.NewNamespacedClient(cl, ns.Name)
 
-	authCfg := deploy.KonnectAPIAuthConfiguration(t, GetCtx(), clientNamespaced,
+	authCfg := deploy.KonnectAPIAuthConfiguration(t, ctx, clientNamespaced,
 		deploy.WithTestIDLabel(testID),
 		func(obj client.Object) {
 			authCfg := obj.(*konnectv1alpha1.KonnectAPIAuthConfiguration)
@@ -54,23 +60,23 @@ func TestKonnectEntities(t *testing.T) {
 		},
 	)
 
-	cp := deploy.KonnectGatewayControlPlane(t, GetCtx(), clientNamespaced, authCfg,
+	cp := deploy.KonnectGatewayControlPlane(t, ctx, clientNamespaced, authCfg,
 		deploy.WithTestIDLabel(testID),
 		deploy.KonnectGatewayControlPlaneLabel(deploy.KonnectTestIDLabel, testID),
 	)
 
-	t.Cleanup(deleteObjectAndWaitForDeletionFn(t, cp.DeepCopy()))
+	t.Cleanup(object.DeleteAndWaitForDeletionFn(context.Background(), t, cl, cp.DeepCopy()))
 
 	t.Logf("Waiting for Konnect ID to be assigned to ControlPlane %s/%s", cp.Namespace, cp.Name)
 	require.EventuallyWithT(t, func(t *assert.CollectT) {
-		err := GetClients().MgrClient.Get(GetCtx(), types.NamespacedName{Name: cp.Name, Namespace: cp.Namespace}, cp)
+		err := cl.Get(ctx, types.NamespacedName{Name: cp.Name, Namespace: cp.Namespace}, cp)
 		require.NoError(t, err)
 		conditions.KonnectEntityIsProgrammed(t, cp)
 	}, testutils.ObjectUpdateTimeout, testutils.ObjectUpdateTick)
 
 	t.Logf("Waiting for ControlPlane and telemetry endpoints to be assigned to ControlPlane %s/%s", cp.Namespace, cp.Name)
 	require.EventuallyWithT(t, func(t *assert.CollectT) {
-		err := GetClients().MgrClient.Get(GetCtx(), types.NamespacedName{Name: cp.Name, Namespace: cp.Namespace}, cp)
+		err := cl.Get(ctx, types.NamespacedName{Name: cp.Name, Namespace: cp.Namespace}, cp)
 		require.NoError(t, err)
 		require.NotEmpty(t, cp.Status.Endpoints)
 		// We will not test the domain name to prevent flakes when the URLs change in Konnect.
@@ -84,7 +90,7 @@ func TestKonnectEntities(t *testing.T) {
 	}, testutils.ObjectUpdateTimeout, testutils.ObjectUpdateTick)
 
 	// Create KongReferenceGrant to allow KongVault (cluster-scoped) to reference KonnectGatewayControlPlane (namespace-scoped).
-	_ = deploy.KongReferenceGrant(t, GetCtx(), clientNamespaced,
+	_ = deploy.KongReferenceGrant(t, ctx, clientNamespaced,
 		deploy.KongReferenceGrantFroms(configurationv1alpha1.ReferenceGrantFrom{
 			Group:     configurationv1alpha1.Group(configurationv1alpha1.GroupVersion.Group),
 			Kind:      "KongVault",
@@ -97,7 +103,7 @@ func TestKonnectEntities(t *testing.T) {
 	)
 
 	t.Run("with Origin ControlPlane", func(t *testing.T) {
-		KonnectEntitiesTestCase(t, konnectEntitiesTestCaseParams{
+		KonnectEntitiesTestCase(t, cl, konnectEntitiesTestCaseParams{
 			cp:     cp,
 			client: clientNamespaced,
 			ns:     ns.Name,
@@ -107,21 +113,21 @@ func TestKonnectEntities(t *testing.T) {
 
 	t.Run("with Mirror ControlPlane", func(t *testing.T) {
 		// Create a Mirror Konnect control plane for the Konnect Entities test case.
-		mirrorCP := deploy.KonnectGatewayControlPlane(t, GetCtx(), clientNamespaced, authCfg,
+		mirrorCP := deploy.KonnectGatewayControlPlane(t, ctx, clientNamespaced, authCfg,
 			deploy.WithTestIDLabel(testID),
 			deploy.WithMirrorSource(cp.GetKonnectID()),
 		)
-		t.Cleanup(deleteObjectAndWaitForDeletionFn(t, mirrorCP.DeepCopy()))
+		t.Cleanup(object.DeleteAndWaitForDeletionFn(context.Background(), t, cl, mirrorCP.DeepCopy()))
 
 		t.Logf("Waiting for Konnect ID to be assigned to ControlPlane %s/%s", mirrorCP.Namespace, mirrorCP.Name)
 		require.EventuallyWithT(t, func(t *assert.CollectT) {
-			err := GetClients().MgrClient.Get(GetCtx(), types.NamespacedName{Name: mirrorCP.Name, Namespace: mirrorCP.Namespace}, mirrorCP)
+			err := cl.Get(ctx, types.NamespacedName{Name: mirrorCP.Name, Namespace: mirrorCP.Namespace}, mirrorCP)
 			require.NoError(t, err)
 			conditions.KonnectEntityIsProgrammed(t, mirrorCP)
 		}, testutils.ObjectUpdateTimeout, testutils.ObjectUpdateTick)
 
 		require.Eventually(t,
-			testutils.ObjectPredicates(t, clients.MgrClient,
+			testutils.ObjectPredicates(t, cl,
 				testutils.MatchCondition[*konnectv1alpha2.KonnectGatewayControlPlane](t).
 					Type(string(konnectv1alpha1.ControlPlaneMirroredConditionType)).
 					Status(metav1.ConditionTrue).
@@ -135,7 +141,7 @@ func TestKonnectEntities(t *testing.T) {
 		require.Equal(t, sdkkonnectcomp.ControlPlaneClusterTypeClusterTypeControlPlane, cp.Status.ClusterType,
 			"status.clusterType must be set to "+sdkkonnectcomp.ControlPlaneClusterTypeClusterTypeControlPlane)
 
-		KonnectEntitiesTestCase(t, konnectEntitiesTestCaseParams{
+		KonnectEntitiesTestCase(t, cl, konnectEntitiesTestCaseParams{
 			cp:     mirrorCP,
 			client: clientNamespaced,
 			ns:     ns.Name,
@@ -151,7 +157,8 @@ type konnectEntitiesTestCaseParams struct {
 	testID string
 }
 
-func KonnectEntitiesTestCase(t *testing.T, params konnectEntitiesTestCaseParams) {
+func KonnectEntitiesTestCase(t *testing.T, cl client.Client, params konnectEntitiesTestCaseParams) {
+	ctx := t.Context()
 	subID := uuid.NewString()[:8]
 	params.testID = params.testID + "-" + subID
 
@@ -174,7 +181,7 @@ func KonnectEntitiesTestCase(t *testing.T, params konnectEntitiesTestCaseParams)
 			}
 		},
 	)
-	t.Cleanup(deleteObjectAndWaitForDeletionFn(t, kr.DeepCopy()))
+	t.Cleanup(object.DeleteAndWaitForDeletionFn(context.Background(), t, cl, kr.DeepCopy()))
 
 	t.Logf("Waiting for KongRoute attached to ControlPlane (serviceless) to be updated with Konnect ID")
 	eventually.KonnectEntityGetsProgrammed(t, ctx, params.client, kr,
@@ -187,7 +194,7 @@ func KonnectEntitiesTestCase(t *testing.T, params konnectEntitiesTestCaseParams)
 
 	t.Log("Making KongRoute service bound")
 	require.EventuallyWithT(t, func(t *assert.CollectT) {
-		err := GetClients().MgrClient.Get(GetCtx(), types.NamespacedName{Name: kr.Name, Namespace: kr.Namespace}, kr)
+		err := cl.Get(ctx, types.NamespacedName{Name: kr.Name, Namespace: kr.Namespace}, kr)
 		require.NoError(t, err)
 		kr.Spec.ServiceRef = &configurationv1alpha1.ServiceRef{
 			Type: configurationv1alpha1.ServiceRefNamespacedRef,
@@ -195,7 +202,7 @@ func KonnectEntitiesTestCase(t *testing.T, params konnectEntitiesTestCaseParams)
 				Name: ks.Name,
 			},
 		}
-		err = GetClients().MgrClient.Update(GetCtx(), kr)
+		err = cl.Update(ctx, kr)
 		require.NoError(t, err)
 	}, testutils.ObjectUpdateTimeout, testutils.ObjectUpdateTick)
 	eventually.KonnectEntityGetsProgrammed(t, ctx, params.client, kr,
@@ -207,10 +214,10 @@ func KonnectEntitiesTestCase(t *testing.T, params konnectEntitiesTestCaseParams)
 
 	t.Log("Making KongRoute serviceless again")
 	require.EventuallyWithT(t, func(t *assert.CollectT) {
-		err := GetClients().MgrClient.Get(GetCtx(), types.NamespacedName{Name: kr.Name, Namespace: kr.Namespace}, kr)
+		err := cl.Get(ctx, types.NamespacedName{Name: kr.Name, Namespace: kr.Namespace}, kr)
 		require.NoError(t, err)
 		kr.Spec.ServiceRef = nil
-		err = GetClients().MgrClient.Update(GetCtx(), kr)
+		err = cl.Update(ctx, kr)
 		require.NoError(t, err)
 	}, testutils.ObjectUpdateTimeout, testutils.ObjectUpdateTick)
 	eventually.KonnectEntityGetsProgrammed(t, ctx, params.client, kr,
@@ -240,13 +247,13 @@ func KonnectEntitiesTestCase(t *testing.T, params konnectEntitiesTestCaseParams)
 			}
 		},
 	)
-	t.Cleanup(deleteObjectAndWaitForDeletionFn(t, kc.DeepCopy()))
+	t.Cleanup(object.DeleteAndWaitForDeletionFn(context.Background(), t, cl, kc.DeepCopy()))
 
 	t.Logf("Waiting for KongConsumer to be updated with Konnect ID and Programmed")
 	eventually.KonnectEntityGetsProgrammed(t, ctx, params.client, kc)
 
 	kp := deploy.ProxyCachePlugin(t, ctx, params.client)
-	t.Cleanup(deleteObjectAndWaitForDeletionFn(t, kp.DeepCopy()))
+	t.Cleanup(object.DeleteAndWaitForDeletionFn(context.Background(), t, cl, kp.DeepCopy()))
 
 	kpb := deploy.KongPluginBinding(t, ctx, params.client,
 		konnect.NewKongPluginBindingBuilder().
@@ -256,7 +263,7 @@ func KonnectEntitiesTestCase(t *testing.T, params konnectEntitiesTestCaseParams)
 			Build(),
 		deploy.WithTestIDLabel(params.testID),
 	)
-	t.Cleanup(deleteObjectAndWaitForDeletionFn(t, kpb.DeepCopy()))
+	t.Cleanup(object.DeleteAndWaitForDeletionFn(context.Background(), t, cl, kpb.DeepCopy()))
 
 	t.Logf("Waiting for KongPluginBinding to be updated with Konnect ID")
 	eventually.KonnectEntityGetsProgrammed(t, ctx, params.client, kpb)
@@ -269,7 +276,7 @@ func KonnectEntitiesTestCase(t *testing.T, params konnectEntitiesTestCaseParams)
 			Build(),
 		deploy.WithTestIDLabel(params.testID),
 	)
-	t.Cleanup(deleteObjectAndWaitForDeletionFn(t, globalKPB.DeepCopy()))
+	t.Cleanup(object.DeleteAndWaitForDeletionFn(context.Background(), t, cl, globalKPB.DeepCopy()))
 
 	t.Logf("Waiting for KongPluginBinding to be updated with Konnect ID")
 	eventually.KonnectEntityGetsProgrammed(t, ctx, params.client, globalKPB)
@@ -284,7 +291,7 @@ func KonnectEntitiesTestCase(t *testing.T, params konnectEntitiesTestCaseParams)
 			kup.Spec.Algorithm = sdkkonnectcomp.UpstreamAlgorithmConsistentHashing.ToPointer()
 		},
 	)
-	t.Cleanup(deleteObjectAndWaitForDeletionFn(t, kup.DeepCopy()))
+	t.Cleanup(object.DeleteAndWaitForDeletionFn(context.Background(), t, cl, kup.DeepCopy()))
 
 	t.Log("Waiting for KongUpstream to be updated with Konnect ID")
 	eventually.KonnectEntityGetsProgrammed(t, ctx, params.client, kup)
@@ -300,23 +307,23 @@ func KonnectEntitiesTestCase(t *testing.T, params konnectEntitiesTestCaseParams)
 
 	t.Log("Waiting for KongTarget to be updated with Konnect ID")
 	require.EventuallyWithT(t, func(t *assert.CollectT) {
-		err := GetClients().MgrClient.Get(GetCtx(), types.NamespacedName{Name: kt.Name, Namespace: params.ns}, kt)
+		err := cl.Get(ctx, types.NamespacedName{Name: kt.Name, Namespace: params.ns}, kt)
 		require.NoError(t, err)
 
 		conditions.KonnectEntityIsProgrammed(t, kt)
 	}, testutils.ObjectUpdateTimeout, testutils.ObjectUpdateTick)
 
 	// Should delete KongTarget because it will block deletion of KongUpstream owning it.
-	t.Cleanup(deleteObjectAndWaitForDeletionFn(t, kt.DeepCopy()))
+	t.Cleanup(object.DeleteAndWaitForDeletionFn(context.Background(), t, cl, kt.DeepCopy()))
 
 	kv := deploy.KongVaultAttachedToCP(t, ctx, params.client, "env", "env-vault", []byte(`{"prefix":"env-vault"}`), params.cp)
 	t.Logf("Waiting for KongVault to be updated with Konnect ID")
 
 	eventually.KonnectEntityGetsProgrammed(t, ctx, params.client, kv)
-	t.Cleanup(deleteObjectAndWaitForDeletionFn(t, kv.DeepCopy()))
+	t.Cleanup(object.DeleteAndWaitForDeletionFn(context.Background(), t, cl, kv.DeepCopy()))
 
 	kcert := deploy.KongCertificateAttachedToCP(t, ctx, params.client, params.cp)
-	t.Cleanup(deleteObjectAndWaitForDeletionFn(t, kcert.DeepCopy()))
+	t.Cleanup(object.DeleteAndWaitForDeletionFn(context.Background(), t, cl, kcert.DeepCopy()))
 
 	t.Logf("Waiting for KongCertificate to get Konnect ID")
 	kcert = eventually.KonnectEntityGetsProgrammed(t, ctx, params.client, kcert)
@@ -328,7 +335,7 @@ func KonnectEntitiesTestCase(t *testing.T, params konnectEntitiesTestCaseParams)
 			ksni.Spec.Name = "test.kong-sni.example.com"
 		},
 	)
-	t.Cleanup(deleteObjectAndWaitForDeletionFn(t, ksni.DeepCopy()))
+	t.Cleanup(object.DeleteAndWaitForDeletionFn(context.Background(), t, cl, ksni.DeepCopy()))
 
 	t.Logf("Waiting for KongSNI to get Konnect ID")
 	eventually.KonnectEntityGetsProgrammed(t, ctx, params.client, ksni,
