+++ test/integration/konnect/suite_test.go
@@ -1,0 +1,11 @@
+package konnect
+
+import (
+	"testing"
+
+	"github.com/kong/kong-operator/v2/test/integration"
+)
+
+func TestMain(m *testing.M) {
+	integration.Suite(m)
+}
